Move MessageFormat out of the account object
