+++ legacy/account/src/main/java/app/k9mail/legacy/account/MessageFormat.kt
@@ -1,0 +1,7 @@
+package app.k9mail.legacy.account
+
+enum class MessageFormat {
+    TEXT,
+    HTML,
+    AUTO,
+}
