+++ legacy/account/src/test/kotlin/app/k9mail/legacy/account/LegacyAccountWrapperTest.kt
@@ -1,6 +1,5 @@
 package app.k9mail.legacy.account
 
-import app.k9mail.legacy.account.Account.MessageFormat
 import app.k9mail.legacy.account.Account.QuoteStyle
 import app.k9mail.legacy.notification.NotificationSettings
 import assertk.assertThat
