+++ legacy/core/src/main/java/com/fsck/k9/AccountPreferenceSerializer.kt
@@ -1,7 +1,6 @@
 package com.fsck.k9
 
 import app.k9mail.legacy.account.Account
-import app.k9mail.legacy.account.Account.MessageFormat
 import app.k9mail.legacy.account.Account.QuoteStyle
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_MAXIMUM_AUTO_DOWNLOAD_MESSAGE_SIZE
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_MESSAGE_FORMAT
@@ -21,6 +20,7 @@ import app.k9mail.legacy.account.DeletePolicy
 import app.k9mail.legacy.account.Expunge
 import app.k9mail.legacy.account.FolderMode
 import app.k9mail.legacy.account.Identity
+import app.k9mail.legacy.account.MessageFormat
 import app.k9mail.legacy.account.ShowPictures
 import app.k9mail.legacy.account.SortType
 import app.k9mail.legacy.account.SpecialFolderSelection
