+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/settings/account/AccountSettingsDataStore.kt
@@ -4,6 +4,9 @@ import androidx.preference.PreferenceDataStore
 import app.k9mail.legacy.account.Account
 import app.k9mail.legacy.account.DeletePolicy
 import app.k9mail.legacy.account.Expunge
+import app.k9mail.legacy.account.LegacyAccount
+import app.k9mail.legacy.account.MessageFormat
+import app.k9mail.legacy.account.QuoteStyle
 import app.k9mail.legacy.account.ShowPictures
 import app.k9mail.legacy.account.SpecialFolderSelection
 import app.k9mail.legacy.notification.NotificationLight
@@ -159,7 +162,7 @@ class AccountSettingsDataStore(
             "expunge_policy" -> account.expungePolicy = Expunge.valueOf(value)
             "max_push_folders" -> account.maxPushFolders = value.toInt()
             "idle_refresh_period" -> account.idleRefreshMinutes = value.toInt()
-            "message_format" -> account.messageFormat = Account.MessageFormat.valueOf(value)
+            "message_format" -> account.messageFormat = MessageFormat.valueOf(value)
             "quote_style" -> account.quoteStyle = Account.QuoteStyle.valueOf(value)
             "account_quote_prefix" -> account.quotePrefix = value
             "account_setup_auto_expand_folder" -> account.autoExpandFolderId = extractFolderId(value)
